certscanner/requirements.txt
elasticsearch==8.13.0
elasticsearch_dsl==8.13.0
nltk==3.8.1
requests==2.31.0
tld==0.13
websocket-client==1.7.0
